refactor(engine): rename JobExecutionHelper to JobExecutorWaitUtils
Improves class naming to better reflect its purpose as a utility class for managing
and waiting on job executor operations in test contexts. The class provides methods
for waiting until jobs are processed and querying job availability.

Ticket: #750
