+++ engine/src/test/java/org/operaton/bpm/engine/test/jobexecutor/SequentialJobAcquisitionTest.java
@@ -16,8 +16,8 @@
  */
 package org.operaton.bpm.engine.test.jobexecutor;
 
-import static org.operaton.bpm.engine.test.util.JobExecutorHelper.waitForJobExecutionRunnablesToFinish;
-import static org.operaton.bpm.engine.test.util.JobExecutorHelper.waitForJobExecutorToProcessAllJobs;
+import static org.operaton.bpm.engine.test.util.JobExecutorWaitUtils.waitForJobExecutionRunnablesToFinish;
+import static org.operaton.bpm.engine.test.util.JobExecutorWaitUtils.waitForJobExecutorToProcessAllJobs;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.text.DateFormat.Field;
